put zod schema in common folder and have client code import zod schema
